backend/requirements.txt
django==4.2.29
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyMySQL==1.1.2
